Move perft.rs to tests
Refactored perft.rs  and moved it into the tests folder.
